# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [productivity]
      summary: List account goals
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [productivity]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [productivity]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [productivity]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /habits:
    get:
      operationId: listHabits
      tags: [productivity]
      summary: List account habits
      responses:
        "200":
          description: Habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [productivity]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [productivity]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [productivity]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /study-items:
    get:
      operationId: listStudyItems
      tags: [productivity]
      summary: List study plan items
      responses:
        "200":
          description: Study items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyItem"
    post:
      operationId: createStudyItem
      tags: [productivity]
      summary: Create a study item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyItemInput"
      responses:
        "201":
          description: Created study item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyItem"
  /study-items/{id}:
    patch:
      operationId: updateStudyItem
      tags: [productivity]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyItemUpdate"
      responses:
        "200":
          description: Updated study item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyItem"
    delete:
      operationId: deleteStudyItem
      tags: [productivity]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /spaces:
    get:
      operationId: listSpaces
      tags: [spaces]
      summary: List task spaces
      responses:
        "200":
          description: Task spaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Space"
    post:
      operationId: createSpace
      tags: [spaces]
      summary: Create a task space
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceInput"
      responses:
        "201":
          description: Created space
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
        "409":
          description: A space with this name already exists
  /spaces/{id}:
    patch:
      operationId: updateSpace
      tags: [spaces]
      summary: Update a task space
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceUpdate"
      responses:
        "200":
          description: Updated space
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
        "404":
          description: Space not found
        "409":
          description: A space with this name already exists
    delete:
      operationId: deleteSpace
      tags: [spaces]
      summary: Delete a task space
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Space not found
...
components:
  schemas:
...
    Goal:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        target: { type: integer }
        current: { type: integer }
        deadline: { type: ["string", "null"], format: date }
        completed: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, target, current, deadline, completed, createdAt, updatedAt]
    GoalInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        target: { type: integer, minimum: 1 }
        deadline: { type: ["string", "null"], format: date }
      required: [title]
    GoalUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        target: { type: integer, minimum: 1 }
        current: { type: integer, minimum: 0 }
        deadline: { type: ["string", "null"], format: date }
        completed: { type: boolean }
    Habit:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        frequency: { type: string, enum: [daily, weekly] }
        streak: { type: integer }
        lastCompleted: { type: ["string", "null"], format: date }
        createdAt: { type: string, format: date-time }
      required: [id, name, frequency, streak, lastCompleted, createdAt]
    HabitInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        frequency: { type: string, enum: [daily, weekly] }
      required: [name]
    HabitUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        frequency: { type: string, enum: [daily, weekly] }
        streak: { type: integer, minimum: 0 }
        lastCompleted: { type: ["string", "null"], format: date }
    StudyItem:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [subject, assignment, exam, review] }
        title: { type: string }
        subject: { type: ["string", "null"] }
        itemDate: { type: ["string", "null"], format: date }
        completed: { type: boolean }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, kind, title, subject, itemDate, completed, notes, createdAt]
    StudyItemInput:
      type: object
      properties:
        kind: { type: string, enum: [subject, assignment, exam, review] }
        title: { type: string, minLength: 1 }
        subject: { type: string }
        itemDate: { type: ["string", "null"], format: date }
        notes: { type: string }
      required: [kind, title]
    StudyItemUpdate:
      type: object
      properties:
        kind: { type: string, enum: [subject, assignment, exam, review] }
        title: { type: string, minLength: 1 }
        subject: { type: ["string", "null"] }
        itemDate: { type: ["string", "null"], format: date }
        completed: { type: boolean }
        notes: { type: ["string", "null"] }
...
    Space:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, color, description, createdAt]
    SpaceInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        description:
          type: string
      required: [name]
    SpaceUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        description:
          type: ["string", "null"]
      required: [name, color]
...