# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  /** @minimum 1 */
  target?: number;
  /** @nullable */
  deadline?: Date | null;
}