# lib/api-zod/src/generated/types/goalUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoalUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minimum 1 */
  target?: number;
  /** @minimum 0 */
  current?: number;
  /** @nullable */
  deadline?: Date | null;
  completed?: boolean;
}