# lib/api-zod/src/generated/types/studyItemInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyItemInputKind } from './studyItemInputKind';

export interface StudyItemInput {
  kind: StudyItemInputKind;
  /** @minLength 1 */
  title: string;
  subject?: string;
  /** @nullable */
  itemDate?: Date | null;
  notes?: string;
}