# lib/api-zod/src/generated/types/studyItemUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyItemUpdateKind } from './studyItemUpdateKind';

export interface StudyItemUpdate {
  kind?: StudyItemUpdateKind;
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  itemDate?: Date | null;
  completed?: boolean;
  /** @nullable */
  notes?: string | null;
}